+++ R/file.R
@@ -71,7 +71,7 @@ file_types <- c(
 #' @param mode A character representation of the mode, in either hexidecimal or symbolic format.
 #' @export
 #' @examples
-#' \dontshow{fs:::pkgdown_tmp("/tmp/filedd4670e2bc60")}
+#' \dontshow{fs:::pkgdown_tmp(c("/tmp/filedd4670e2bc60", "/tmp/file87bb1f8fcb02"))}
 #' x <- file_create(file_temp(), "000")
 #' file_chmod(x, "777")
 #' file_info(x)$permissions
@@ -84,9 +84,16 @@ file_types <- c(
 #'
 #' file_chmod(x, "u+wr")
 #' file_info(x)$permissions
+#'
+#' # It is also vectorized
+#' files <- c(x, file_create(file_temp(), "000"))
+#' file_chmod(files, "a+rwx")
+#' file_info(files)$permissions
+#'
+#' file_chmod(files, c("644", "600"))
+#' file_info(files)$permissions
 file_chmod <- function(path, mode) {
   assert_no_missing(path)
-  stopifnot(length(mode) == 1)
   mode <- as_fs_perms(mode, mode = file_info(path)$permissions)
 
   old <- path_expand(path)